main_window: Use a default value for open_filechooser
Callback is a callback function to open a safe. It can be called from
an action activate event or a clicked event of a button. In the latter
case, no user_data is set. Having None as a default value allows to
handle both events reliably.

As a side-note, it will make it easier to port main_window to a
Gtk.Template.